+++ SQLSkyline/Helper.cs
@@ -574,10 +574,10 @@ namespace prefSQL.SQLSkyline
             return dataTableRowList.Select(dataRow => dataRow.ItemArray).ToList();
         }
 
-        public static Dictionary<int, object[]> GetDictionaryFromDataTable(DataTable dataTable, int uniqueIdColumnIndex)
+        public static Dictionary<long, object[]> GetDictionaryFromDataTable(DataTable dataTable, int uniqueIdColumnIndex)
         {
             List<object[]> objectArrayFromDataTableOrig = GetObjectArrayFromDataTable(dataTable);
-            return objectArrayFromDataTableOrig.ToDictionary(dataRow => (int)dataRow[uniqueIdColumnIndex]);
+            return objectArrayFromDataTableOrig.ToDictionary(dataRow => (long)dataRow[uniqueIdColumnIndex]);
         }
 
         public static DataTable GetDataTableFromSQL(string strQuery, string strConnection, string strProvider)
